Journal Entry Controller Completed by Using authentication method

+++ journalApp/src/main/java/com/sawan/journalApp/service/JournalEntryService.java
@@ -29,7 +29,7 @@ public class JournalEntryService {
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         } catch (Exception e) {
             log.error("Exception ", e);
             throw new RuntimeException("An error occured while saving the entry", e);
@@ -48,15 +48,22 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName) {
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed=false;
         try {
             User user = userService.findByUserName(userName);
-            user.getJournalEntries().removeIf(X -> X.getId().equals(id));
-            userService.saveEntry(user);
-            journalEntryRepository.deleteById(id);
+            removed = user.getJournalEntries().removeIf(X -> X.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+
         } catch (Exception e) {
             log.error("Delete error", e);
+            throw new RuntimeException("An Error occurred while delete the entry", e);
         }
+        return removed;
 
     }
 }
